backend/requirements.txt
networkx
typing
google-generativeai
numpy
dotenv
flask
pypdf
re
spacy
fitz
psycopg2-binary